test(solid-query/useQuery): move type-only test to 'useQuery.test-d.tsx'
